+++ dask_geomodeling/utils.py
@@ -3,6 +3,7 @@ import math
 import pytz
 import os
 import warnings
+from distutils.version import LooseVersion
 from functools import lru_cache
 from itertools import repeat
 
@@ -18,6 +19,9 @@ from shapely.geometry import box, Point
 from shapely import wkb as shapely_wkb
 
 import fiona
+import fiona.crs
+import geopandas
+
 
 POLYGON = "POLYGON (({0} {1},{2} {1},{2} {3},{0} {3},{0} {1}))"
 
@@ -25,6 +29,10 @@ try:
     from fiona import Env as fiona_env  # NOQA
 except ImportError:
     from fiona import drivers as fiona_env  # NOQA
+
+GEOPANDAS_GTE_0_7_0 = LooseVersion(geopandas.__version__) >= LooseVersion("0.7.0")
+if GEOPANDAS_GTE_0_7_0:
+    from pyproj import CRS
 
 
 def get_index(values, no_data_value):
@@ -334,9 +342,17 @@ def get_sr(user_input):
 
 
 def get_crs(user_input):
+    """Return CRS for user input.
+
+    Args:
+      user_input (str): a WKT, PROJ4, or EPSG:xxxx crs representation
+
+    Returns:
+      for geopandas >= 0.7: pyproj.CRS
+      for geopandas < 0.7: dict
     """
-    Return fiona CRS dictionary for user input.
-    """
+    if GEOPANDAS_GTE_0_7_0:
+        return CRS(user_input)
     wkt = osr.GetUserInputAsWKT(str(user_input))
     sr = osr.SpatialReference(wkt)
     key = str("GEOGCS") if sr.IsGeographic() else str("PROJCS")
@@ -352,10 +368,21 @@ def get_crs(user_input):
 
 def crs_to_srs(crs):
     """
-    Recover our own WKT definition of projections from a fiona CRS
+    Recover our own WKT definition of projections from a pyproj / fiona CRS
+
+    Args:
+      crs (pyproj.CRS or dict): what is returned from GeoDataFrame().crs
+
+    Returns:
+      string: a "EPSG:xxx" or WKT representation of the crs.
     """
-    proj4_str = fiona.crs.to_string(crs)
-    return get_epsg_or_wkt(proj4_str)
+    if crs is None:
+        return
+    elif isinstance(crs, dict):  # geopandas < 0.7
+        proj4_str = fiona.crs.to_string(crs)
+        return get_epsg_or_wkt(proj4_str)
+    else:  # geopandas >= 0.7
+        return crs.to_string()
 
 
 def wkb_transform(wkb, src_sr, dst_sr):